requirements.txt
selenium==4.19.0
pytest==8.1.1
allure-pytest==2.13.5
requests==2.31.0
pytest-xdist==3.3.1
Faker==28.1.0
